ex4 - Neural Network Learning
In the first part of the exercise, we implement the backpropagation algorithm to learn the parameters for the neural network.
In the second  part of the exercise, we implement the backpropagation algorithm to compute the gradient for the neural network cost function. we need to complete the nnCostFunction.m so that it returns an appropriate value for grad. Once we have computed the gradient, we will be able to train the neural network by minimizing the cost function J(Θ) using an advanced optimizer such as fmincg.